+++ src/main/java/com/showcle/global/model/FileModel.java
@@ -1,6 +1,5 @@
 package com.showcle.global.model;
 
-import com.showcle.global.enums.FileType;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
@@ -12,7 +11,7 @@ import lombok.ToString;
 @Builder
 public class FileModel {
 
-    private int idx;
+    private long idx;
     private int ownerIdx;
     private String type;
     private String fileType;
